+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+Master
+-------------------------
+
+* Interface changes
+    * Remove exponential incrementer
+    * Rename data! to json!
+    * Rename increments to steps & change default to 10
+    * Remove linear
+    * Add min_size
+    * Add step_size
 
 0.1.0 (Current release)
 -------------------------